orders work + insert some mapping

+++ Project2_backend/src/main/java/com/revature/Project2_backend/controllers/OrderResource.java
@@ -4,6 +4,7 @@ import com.revature.Project2_backend.model.forUser.User;
 import com.revature.Project2_backend.service.OrdersService;
 import com.revature.Project2_backend.service.UserService;
 import lombok.AllArgsConstructor;
+import org.apache.coyote.Response;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,22 +21,20 @@ public class OrderResource {
   private final UserService userService;
 
 
-  @GetMapping
+  @GetMapping("/getOrderList")
   public ResponseEntity<List<Orders>> getOrders(){
     List<Orders> orders = ordersService.findAllOrders();
     return new ResponseEntity<>(orders, HttpStatus.OK);
   }
 
-  @PostMapping
+  @PostMapping("/addOrders")
   public ResponseEntity<Orders> addOrder(@RequestBody Orders order){
     User user = userService.findUserById(order.getUserId());
-    Orders ordered;
     if(user != null){
-      ordered = ordersService.addOrder(order);
+      return new ResponseEntity<>(ordersService.addOrder(order), HttpStatus.OK);
     }
-    else ordered = null;
-
-    return new ResponseEntity<>(ordered, HttpStatus.CREATED);
+    else
+      return new ResponseEntity<>(HttpStatus.NOT_FOUND);
   }
 
   @GetMapping("/orderId/{orderId}")
@@ -44,15 +43,22 @@ public class OrderResource {
     return new ResponseEntity<>(order, HttpStatus.OK);
   }
 
-  @PutMapping
+  @PutMapping("/orderUpdate")
   public ResponseEntity<Orders> updateOrder(@RequestBody Orders upOrder){
-    Orders updatedOrder = ordersService.updateOrder(upOrder);
-    return new ResponseEntity<>(updatedOrder, HttpStatus.OK);
+    Orders findOrder = ordersService.findByOrderId(upOrder.getOrderId());
+    if(findOrder != null){
+      return new ResponseEntity<>(ordersService.updateOrder(upOrder), HttpStatus.OK);
+    }
+    else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
   }
   @Transactional
   @DeleteMapping("/orderId/{orderId}")
   public ResponseEntity<?> deleteOrder(@PathVariable("orderId") long orderId) {
-    ordersService.deleteOrders(orderId);
-    return new ResponseEntity<>(HttpStatus.OK);
+    Orders findOrder = ordersService.findByOrderId(orderId);
+    if(findOrder != null){
+      ordersService.deleteOrders(orderId);
+      return new ResponseEntity<>(HttpStatus.OK);
+    }
+    else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
   }
 }
